+++ Atlas.Domain/Promo.cs
@@ -15,7 +15,7 @@ namespace Atlas.Domain
 
         public int DiscountPrice { get; set; }
 
-        public int DiscountPercent { get; set; }
+        public float DiscountPercent { get; set; }
 
         public bool ForAllGoods { get; set; }
